complete feature set but doesnt compile

+++ src/mm_toolbox/moving_average/base.pxd
@@ -3,13 +3,13 @@ from mm_toolbox.ringbuffer.onedim cimport RingBufferOneDim
 
 cdef class MovingAverage:
     cdef:
-        Py_ssize_t          _window
+        int                 _window
         bint                _is_fast
         bint                _is_warm
         double              _value
         RingBufferOneDim    _values
 
-    # def void              __init__(self, Py_ssize_t window, bint is_fast)
+    # def void              __init__(self, int window, bint is_fast)
     cdef inline void        ensure_warm(self)
     cdef inline void        ensure_not_fast(self)
     cdef inline void        push_to_ringbuffer(self)
@@ -20,7 +20,7 @@ cdef class MovingAverage:
     cpdef double            get_value(self)
     cpdef cnp.ndarray       get_values(self)
 
-    # def Py_ssize_t        __len__(self)
+    # def int               __len__(self)
     # def bool              __contains__(self, double value)
     # def generator         __iter__(self)
-    # def double            __getitem__(self, Py_ssize_t idx)
+    # def double            __getitem__(self, int idx)
